Small review changes (changed Postgress model to take in bytes instead and therefore had to update segments to use byts)

+++ src/main/java/segmentgenerator/SegmentGenerator.java
@@ -74,7 +74,7 @@ public class SegmentGenerator {
         ValueCompressionModel valueModel = compressionModel.getValueCompressionModel();
         TimeStampCompressionModel timeStampModel = compressionModel.getTimeStampCompressionModel();
 
-        return new Segment(this.timeSeriesId, startTime, endTime, valueModel.getValueCompressionModelType().ordinal(), valueModel.getBlobRepresentation(), timeStampModel.getTimeStampCompressionModelType().ordinal(), timeStampModel.getBlobRepresentation());
+        return new Segment(this.timeSeriesId, startTime, endTime, (byte) valueModel.getValueCompressionModelType().ordinal(), valueModel.getBlobRepresentation(), (byte) timeStampModel.getTimeStampCompressionModelType().ordinal(), timeStampModel.getBlobRepresentation());
     }
 
 }
